test: keep operator settings fixture within package topology
